Rename decryptor/padgen to decryptor/decryptor
