+++ source/decryptor/decryptor.h
